renames index.js to app.js and adds socket.io documentation
