--- index.js
@@ -1,23 +1,0 @@
-const express = require("express");
-const bodyParser = require("body-parser");
-const app = express();
-const port = 3000;
-const db = require("./queries");
-
-app.use(bodyParser.json());
-app.use(
-  bodyParser.urlencoded({
-    extended: true,
-  })
-);
-
-app.get("/", (request, response) => {
-  response.json({ info: "Node.js, Express, and Postgres API" });
-});
-
-app.listen(port, () => {
-  console.log(`App running on port ${port}.`);
-});
-
-app.get("/messages", db.getMessages);
-app.post("/messages", db.createMessage);
